move hydrology methods to base class from space
