+++ landlab/components/erosion_deposition/generalized_erosion_deposition.py
@@ -1,4 +1,6 @@
 import numpy as np
+from six import string_types
+
 from landlab import Component
 from landlab import RasterModelGrid
 from landlab.utils.return_array import return_array_at_node
@@ -138,3 +140,118 @@ class _GeneralizedErosionDeposition(Component):
         self.slope[:] = ((self.topographic__elevation -
                           self.topographic__elevation[self.flow_receivers]) /
                          self.link_lengths[self.link_to_reciever])
+
+
+
+    #three choices for erosion methods:
+    def simple_stream_power(self):
+        """Use non-threshold stream power.
+
+        simple_stream_power uses no entrainment or erosion thresholds,
+        and uses either q=A^m or q=Q^m depending on discharge method. If
+        discharge method is None, default is q=A^m.
+        """
+        if self.discharge_method == None:
+            self.Q_to_the_m[:] = np.power(self.grid.at_node['drainage_area'], self.m_sp)
+        else:
+            if self.discharge_method == 'area_field':
+                if self.area_field is not None:
+                    if isinstance(self.area_field, string_types):
+                        self.drainage_area = self._grid.at_node[self.area_field]
+                    elif len(self.area_field) == self.grid.number_of_nodes:
+                        self.drainage_area = np.array(self.area_field)
+                    else:
+                        raise TypeError('Supplied type of area_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+                self.Q_to_the_m[:] = np.power(self.drainage_area, self.m_sp)
+            elif self.discharge_method == 'discharge_field':
+                if self.discharge_field is not None:
+                    if isinstance(self.discharge_field, string_types):
+                        self.q[:] = self._grid.at_node[self.discharge_field]
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    elif len(self.discharge_field) == self.grid.number_of_nodes:
+                        self.q[:] = np.array(self.discharge_field)
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    else:
+                        raise TypeError('Supplied type of discharge_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+            else:
+                raise TypeError('Unrecognized discharge_method')
+
+        #TODO: FIGURE OUT WHY WE BOTHER TO CALC AND STORE BOTH ES/ER AND
+        #**_EROSION_TERM?
+
+
+    def threshold_stream_power(self):
+        """Use stream power with entrainment/erosion thresholds.
+
+        threshold_stream_power works the same way as simple SP but includes
+        user-defined thresholds for sediment entrainment and bedrock erosion.
+        """
+        self.Q_to_the_m = np.zeros(len(self.grid.at_node['drainage_area']))
+        if self.method == 'threshold_stream_power' and self.discharge_method == None:
+            self.Q_to_the_m[:] = np.power(self.grid.at_node['drainage_area'], self.m_sp)
+        elif self.method == 'threshold_stream_power' and self.discharge_method is not None:
+            if self.discharge_method == 'drainage_area':
+                if self.area_field is not None:
+                    if isinstance(self.area_field, string_types):
+                        self.drainage_area = self._grid.at_node[self.area_field]
+                    elif len(self.area_field) == self.grid.number_of_nodes:
+                        self.drainage_area = np.array(self.area_field)
+                    else:
+                        raise TypeError('Supplied type of area_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+                self.Q_to_the_m[:] = np.power(self.drainage_area, self.m_sp)
+            elif self.discharge_method == 'discharge_field':
+                if self.discharge_field is not None:
+                    if isinstance(self.discharge_field, string_types):
+                        self.q[:] = self._grid.at_node[self.discharge_field]
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    elif len(self.discharge_field) == self.grid.number_of_nodes:
+                        self.q[:] = np.array(self.discharge_field)
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    else:
+                        raise TypeError('Supplied type of discharge_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+
+
+    def stochastic_hydrology(self):
+        """Allows custom area and discharge fields, no default behavior.
+
+        stochastic_hydrology forces the user to supply either an array or
+        field name for either drainage area or discharge, and will not
+        default to q=A^m.
+        """
+        self.Q_to_the_m = np.zeros(len(self.grid.at_node['drainage_area']))
+        if self.method == 'stochastic_hydrology' and self.discharge_method == None:
+            raise TypeError('Supply a discharge method to use stoc. hydro!')
+        elif self.discharge_method is not None:
+            if self.discharge_method == 'drainage_area':
+                if self.area_field is not None:
+                    if isinstance(self.area_field, string_types):
+                        self.drainage_area = self._grid.at_node[self.area_field]
+                    elif len(self.area_field) == self.grid.number_of_nodes:
+                        self.drainage_area = np.array(self.area_field)
+                    else:
+                        raise TypeError('Supplied type of area_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+                self.Q_to_the_m[:] = np.power(self.grid.at_node['drainage_area'], self.m_sp)
+            elif self.discharge_method == 'discharge_field':
+                if self.discharge_field is not None:
+                    if isinstance(self.discharge_field, string_types):
+                        self.q[:] = self._grid.at_node[self.discharge_field]
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    elif len(self.discharge_field) == self.grid.number_of_nodes:
+                        self.q[:] = np.array(self.discharge_field)
+                        self.Q_to_the_m[:] = np.power(self.q, self.m_sp)
+                    else:
+                        raise TypeError('Supplied type of discharge_field ' +
+                                'was not recognised, or array was ' +
+                                'not nnodes long!')
+            else:
+                raise ValueError('Specify discharge method for stoch hydro!')
